Reorg upload validators and shift code to parent class
There was some duplicated code in the upload validators that has been DRY'ed and shifted to the parent class, which has been renamed to be more explicit to `BaseUploadValidator`

As part of this reorg the bug where the filename for an upload with virus was showing as empty quotes has been fixed.
