+++ app/validators/metadata_presenter/max_size_validator.rb
@@ -1,5 +1,5 @@
 module MetadataPresenter
-  class MaxSizeValidator < UploadValidator
+  class MaxSizeValidator < BaseUploadValidator
     def error_name
       'invalid.too-large'
     end
