+++ app/validators/metadata_presenter/multiupload_validator.rb
@@ -1,20 +1,7 @@
 module MetadataPresenter
-  class MultiuploadValidator < BaseValidator
+  class MultiuploadValidator < BaseUploadValidator
     def invalid_answer?
       user_answer.errors.any? { |error| error.attribute.to_s == error_name }
-    end
-
-    def user_answer
-      page_answers.uploaded_files.find do |uploaded_file|
-        component.id == uploaded_file.component.id
-      end
-    end
-
-    def error_message_hash
-      {
-        control: page_answers.send(component.id)[component.id].last['original_filename'],
-        schema_key.to_sym => component.validation[schema_key]
-      }
     end
 
     def error_name
